+++ Rockets2D/src/classes/scene/GameElements/FreeBodyDiagram.cpp
@@ -5,10 +5,10 @@ FreeBodyDiagram::FreeBodyDiagram(sf::RenderTexture* t, sf::RenderWindow* w, int 
 	fbdPtr = ptr;
 	canDraw = canD;
 	size = s;
-	body.setRadius(size / 10);
+	body.setRadius((float)size / 10);
 	body.setPointCount(50);
 	body.setOrigin(sf::Vector2f(body.getLocalBounds().height / 2, body.getLocalBounds().height / 2));
-	body.setPosition(sf::Vector2f(x, y));
+	body.setPosition(sf::Vector2f((float)x, (float)y));
 	body.setFillColor(sf::Color::Black);
 	body.setOutlineThickness(2);
 	body.setOutlineColor(sf::Color(0, 255, 65));
@@ -38,7 +38,7 @@ void FreeBodyDiagram::forceToArrow() {
 		arrows.clear();
 		for (Vector2 force : fbdPtr->getForces()) {
 			if (force.getMagnitude() > 0) {
-				arrows.push_back(Arrow(force.getMagnitude() / maxForce * ((long double)size - (long double)10), 10, x, y, (-force.getAngle() * 180 / M_PI) - 90));
+				arrows.push_back(Arrow((int)(force.getMagnitude() / maxForce * ((long double)size - (long double)10)), 10, x, y, (-force.getAngle() * 180 / M_PI) - 90));
 				arrows.back().setDirection(force.getIsPush());
 			}
 		}
