+++ Rockets2D/src/classes/scene/Misc/Arrow.cpp
@@ -7,31 +7,31 @@ Arrow::Arrow(int l, int w, int xP, int yP, long double a) {
 	y = yP;
 	angle = a;
 	arrowHead.setPointCount(3);
-	arrowHead.setRadius(width / 2);
+	arrowHead.setRadius((float)width / 2);
 	arrowHead.setFillColor(sf::Color::Black);
-	arrowHead.setOrigin(arrowHead.getLocalBounds().width / 2, length);
-	arrowHead.setPosition(x, y);
-	arrowHead.setRotation(a);
+	arrowHead.setOrigin(arrowHead.getLocalBounds().width / 2, (float)length);
+	arrowHead.setPosition((float)x, (float)y);
+	arrowHead.setRotation((float)a);
 	arrowHead.setOutlineThickness(2);
 	arrowHead.setOutlineColor(sf::Color(0, 255, 65));
-	body.setSize(sf::Vector2f(2, length));
-	body.setOrigin(body.getLocalBounds().width / 2, length);
+	body.setSize(sf::Vector2f(2, (float)length));
+	body.setOrigin(body.getLocalBounds().width / 2, (float)length);
 	body.setFillColor(sf::Color(0, 255, 65));
-	body.setPosition(x, y);
-	body.setRotation(a);
+	body.setPosition((float)x, (float)y);
+	body.setRotation((float)a);
 }
 
 void Arrow::setLength(int l) {
 	length = l;
-	arrowHead.setOrigin(arrowHead.getLocalBounds().width / 2, length);
-	body.setSize(sf::Vector2f(2, length));
-	body.setOrigin(body.getLocalBounds().width / 2, length);
+	arrowHead.setOrigin(arrowHead.getLocalBounds().width / 2, (float)length);
+	body.setSize(sf::Vector2f(2, (float)length));
+	body.setOrigin(body.getLocalBounds().width / 2, (float)length);
 }
 
 void Arrow::setAngle(int a) {
 	angle = a;
-	arrowHead.setRotation(a);
-	body.setRotation(a);
+	arrowHead.setRotation((float)a);
+	body.setRotation((float)a);
 }
 
 void Arrow::draw(sf::RenderTexture* rt) {
@@ -46,6 +46,6 @@ void Arrow::setDirection(bool d) {
 		arrowHead.setRadius(-arrowHead.getRadius());
 	} else {
 		arrowHead.setOrigin(arrowHead.getLocalBounds().width / 2 - 2, length - arrowHead.getRadius());
-		setAngle(angle+180);
+		setAngle((int)angle+180);
 	}
 }
